+++ SocialNetwork/client/src/app/components/header/header.component.css
