+++ SocialNetwork/client/src/app/components/header/header.component.ts
@@ -1,0 +1,26 @@
+import { Component, OnInit, Input } from '@angular/core';
+// import { BreakpointObserver, Breakpoints, BreakpointState } from '@angular/cdk/layout';
+// import { Observable } from 'rxjs';
+// import { map } from 'rxjs/operators';
+import { ThemePalette } from '@angular/material/core';
+
+
+@Component({
+  selector: 'app-header',
+  templateUrl: './header.component.html',
+  styleUrls: ['./header.component.css']
+})
+
+export class HeaderComponent {
+
+  @Input()
+  color: ThemePalette;
+
+  // isHandset$: Observable<boolean> = this.breakpointObserver.observe(Breakpoints.Handset)
+  //   .pipe(
+  //     map(result => result.matches)
+  //   );
+
+  // constructor(private breakpointObserver: BreakpointObserver) { }
+  constructor() { }
+}
