+++ SocialNetwork/client/src/app/components/profile-page/user-info/user-info.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { Profile } from '../models/profile';
+import { Profile } from '../../../models/profile';
 
 @Component({
   selector: 'app-user-info',
