+++ SocialNetwork/client/src/app/components/profile-page/user-page/user-page.component.scss
