+++ SocialNetwork/client/src/app/components/profile-page/user-tabs/user-tabs.component.scss
@@ -1,0 +1,9 @@
+.example-form {
+  min-width: 150px;
+  max-width: 500px;
+  width: 100%;
+}
+
+.example-full-width {
+  width: 100%;
+}
