+++ SocialNetwork/client/src/app/components/registration-form/registration-form.component.css
