+++ SocialNetwork/client/src/app/models/post.ts
@@ -1,8 +1,8 @@
-import { Profile } from "./profile";
+import { Profile } from './profile';
 
-export class Post {
-    id: number;
-    text: string;
-    datetime: string;
-    profile: Profile;
-  }
+export interface Post {
+  id: number;
+  text: string;
+  datetime: string;
+  profile: Profile;
+}
