+++ SocialNetwork/client/src/app/models/user.ts
@@ -2,7 +2,7 @@ export class User {
     constructor(
       public login?: string,
       public email?: string,
-      public password?: string, 
+      public password?: string,
       public firstname?: string,
       public lastname?: string,
     ) {}
