+++ SocialNetwork/client/src/app/modules/material/material.module.spec.ts
@@ -1,0 +1,13 @@
+import { MaterialModule } from './material.module';
+
+describe('MaterialModule', () => {
+  let materialModule: MaterialModule;
+
+  beforeEach(() => {
+    materialModule = new MaterialModule();
+  });
+
+  it('should create an instance', () => {
+    expect(materialModule).toBeTruthy();
+  });
+});
