+++ SocialNetwork/client/src/app/services/app.component.html
