+++ SocialNetwork/client/src/app/services/app.component.scss
