+++ SocialNetwork/client/src/app/services/app.component.spec.ts
