+++ SocialNetwork/client/src/app/services/model-services/profile.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Profile } from './models/profile';
+import { Profile } from '../../models/profile';
 import { Observable, of } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
 
@@ -8,13 +8,13 @@ import { HttpClient } from '@angular/common/http';
 })
 export class ProfileService {
   getProfile(login: string): Observable<Profile> {
-    return this.http.get<Profile>("http://localhost:5000/api/profiles/login/?login=" + login);
+    return this.http.get<Profile>('http://localhost:5000/api/profiles/login/?login=' + login);
   }
   getSubscribers(profileId: number): Observable<Profile[]> {
-    return this.http.get<Profile[]>("http://localhost:5000/api/followings/subscribers/?id=" + profileId.toString());
+    return this.http.get<Profile[]>('http://localhost:5000/api/followings/subscribers/?id=' + profileId.toString());
   }
   getBloggers(profileId: number): Observable<Profile[]> {
-    return this.http.get<Profile[]>("http://localhost:5000/api/followings/bloggers/?id=" + profileId.toString());
+    return this.http.get<Profile[]>('http://localhost:5000/api/followings/bloggers/?id=' + profileId.toString());
   }
   constructor(private http: HttpClient) { }
 }
