+++ SocialNetwork/client/src/app/services/validators/input-data-validator.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 
-import {FormControl, Validators} from '@angular/forms';
+import { FormControl, Validators } from '@angular/forms';
 
 @Injectable({
   providedIn: 'root'
@@ -12,33 +12,33 @@ export class InputDataValidatorService {
 
   constructor() { }
 
-  getLoginValidator(){
+  getLoginValidator() {
     return this.loginValidator;
   }
 
-  getEmailValidator(){
+  getEmailValidator() {
     return this.emailValidator;
   }
 
-  getPasswordValidator(){
+  getPasswordValidator() {
     return this.passwordValidator;
   }
 
   getEmailErrorMessage() {
     return this.emailValidator.hasError('required') ? 'You must enter a value' :
-        this.emailValidator.hasError('email') ? 'Not a valid email' :
-            '';
+      this.emailValidator.hasError('email') ? 'Not a valid email' :
+        '';
   }
 
   getLoginErrorMessage() {
     return this.loginValidator.hasError('required') ? 'You must enter a login value' :
-        this.loginValidator.hasError('minlength') ? 'You have to enter more than 6 symbols' :
-            '';
+      this.loginValidator.hasError('minlength') ? 'You have to enter more than 6 symbols' :
+        '';
   }
 
   getPasswordErrorMessage() {
     return this.passwordValidator.hasError('required') ? 'You must enter a password value' :
-        this.passwordValidator.hasError('minlength') ? 'You have to enter more than 8 symbols' :
-            '';
+      this.passwordValidator.hasError('minlength') ? 'You have to enter more than 8 symbols' :
+        '';
   }
 }
